Moved file coin related code into own directory
